Unify the functor definition of elementwise add, sub, mul, div, floordiv, max, min. (#35684)
