+++ paddle/fluid/operators/elementwise/elementwise_floordiv_op.cu
@@ -11,24 +11,12 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
+
 #include "paddle/fluid/operators/elementwise/elementwise_floordiv_op.h"
 #include "paddle/fluid/operators/elementwise/elementwise_op_broadcast.cu.h"
 
-namespace ops = paddle::operators;
-namespace plat = paddle::platform;
-
 namespace paddle {
 namespace operators {
-
-template <typename T>
-struct CudaFloorDivFunctor {
-  inline HOSTDEVICE T operator()(const T argv[]) const {
-    PADDLE_ENFORCE(argv[1] != 0,
-                   "InvalidArgument: divide by zero "
-                   "encountered in floor-divide ops, please check.\n");
-    return static_cast<T>(std::trunc(argv[0] / argv[1]));
-  }
-};
 
 template <typename T>
 class ElementwiseFloorDivKernel<platform::CUDADeviceContext, T>
@@ -42,12 +30,15 @@ class ElementwiseFloorDivKernel<platform::CUDADeviceContext, T>
 
     int axis = PackTensorsIntoVector<T>(ctx, &ins, &outs);
     LaunchElementwiseCudaKernel<ElementwiseType::kBinary, T, T>(
-        cuda_ctx, ins, &outs, axis, CudaFloorDivFunctor<T>());
+        cuda_ctx, ins, &outs, axis, FloorDivFunctor<T>());
   }
 };
 
 }  // namespace operators
 }  // namespace paddle
+
+namespace ops = paddle::operators;
+namespace plat = paddle::platform;
 
 REGISTER_OP_CUDA_KERNEL(
     elementwise_floordiv,
